Chapter 5 Assignment
Due Nov. 6, 2016.
Submitted Nov. 3, 2016.

Write a function areaOfCircle(r) which returns the area of a circle of radius r

As a refresher, the area of any circle is equal to the radius squared, multiplied by pi (where pi is 3.14159....).

Don’t forget to include the math module, where pi is defined.